Implement settings class and break app into modules
