+++ requirements.txt
@@ -1,5 +1,6 @@
 fastapi
 pony
 requests
+toml
 urlpath
 uvicorn
