+++ src/config/__init__.py
@@ -1,0 +1,3 @@
+from . import _settings
+
+settings = _settings.Settings.load()
