+++ src/config/settings.toml
@@ -1,0 +1,3 @@
+series_dirs = [
+    "/home/anne/manga/"   
+]
